product all child delete

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -96,6 +97,13 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $variant = ProductVariant::findOrFail($id);
+        $variantItem = ProductVariantItem::where('product_variant_id', $variant->id)->count();
+        if($variantItem > 0){
+            return response()->json([
+                'message' => 'You can not delete it. It has item. Please delete first item',
+                'status' => 'error'
+               ]);
+        }
         $variant->delete();
 
         return response()->json([
